refactor: move Git::Repository::Factories and PathResolver to Git namespace
Both modules were nested under Git::Repository:: but were never mixed into
Git::Repository instances. They are extended onto / used by the Git module
directly, matching the existing Git::Configuring precedent.

- Git::Repository::Factories -> Git::Factories (lib/git/factories.rb)
- Git::Repository::PathResolver -> Git::PathResolver (lib/git/path_resolver.rb)
- Update lib/git.rb: require + extend Git::Factories
- Update lib/git/repository.rb: remove require git/repository/factories
- Move and update spec files to spec/unit/git/
