+++ lib/git/repository.rb
@@ -9,7 +9,6 @@ require 'git/repository/branching'
 require 'git/repository/context_helpers'
 require 'git/repository/committing'
 require 'git/repository/diffing'
-require 'git/repository/factories'
 require 'git/repository/inspecting'
 require 'git/repository/logging'
 require 'git/repository/maintenance'
